Let browser handle downloads directly - part 2(2)
dCacheView is limited to file-by-file downloads, which is frustrating
users since there is support for multi-file selection.

This patch builds on the previous one to enable downloads for multi-file
selections. The simplest solution is implemented that triggers multiple
file downloads in the browser, this is also found to work best with
tablets and smartphones, where the traditional method by handling
multi-file download by providing a .zip archive of the files is really
cumbersome to handle.

There is also a completion of the implementation to handle
sub-directories in the shared-files view and small fixes to part 1 noted
in review that was easiest to squash into this commit.

The end result is that it is now possible to select multiple files in a
directory and then download them in a smooth manner.

Fixes: https://github.com/dCache/dcache-view/issues/268

Signed-off-by: Niklas Edmundsson <nikke@hpc2n.umu.se>